Split linuwux.c into modules under src/modules/
linuwux.c had grown into a single ~850-line file covering CPUID
spoofing, SIGSYS redirect, HwProfileGuid, faketime, and the
sigaction()/prctl() interposition all at once. Split by concern into
src/modules/ (cpuid, sigsys, registry, faketime, hooks, common),
leaving linuwux.c as just the constructor that ties them together.

Splitting turns the cross-module functions into external-linkage
symbols, which by default become exported from the .so -- build now
uses -fvisibility=hidden with explicit default-visibility only on the
four real interposition targets (sigaction, prctl, clock_gettime,
gettimeofday), so the exported surface is unchanged.

build.sh, CI, and README updated for the new layout.
