+++ src/modules/registry.h
@@ -1,0 +1,26 @@
+/*
+ * Copyright (C) 2026 brcly
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published
+ * by the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ *
+ * SPDX-License-Identifier: AGPL-3.0-or-later
+ */
+
+#ifndef LINUWUX_REGISTRY_H
+#define LINUWUX_REGISTRY_H
+
+/* Write Hardware Profiles\0001\HwProfileGuid once (arm leaf). */
+void linuwux_set_hwprofile_guid(void);
+
+#endif /* LINUWUX_REGISTRY_H */
